+++ R/LearnerClassifKKNN.R
@@ -1,19 +1,13 @@
 #' @title k-Nearest-Neighbor Classification Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_classif.kknn
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerClassif].
-#'
-#' @section Construction:
-#' ```
-#' LearnerClassifKKNN$new()
-#' mlr3::mlr_learners$get("classif.kknn")
-#' mlr3::lrn("classif.kknn")
-#' ```
+#' @name mlr_learners_classif.kknn
 #'
 #' @description
 #' k-Nearest-Neighbor classification.
 #' Calls [kknn::kknn()] from package \CRANpkg{kknn}.
+#'
+#' @templateVar id classif.kknn
+#' @template section_dictionary_learner
 #'
 #' @references
 #' \cite{mlr3learners}{hechenbichler_2004}
@@ -24,10 +18,12 @@
 #'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name classif.kknn
 #' @template example
 LearnerClassifKKNN = R6Class("LearnerClassifKKNN", inherit = LearnerClassif,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ParamSet$new(list(
         ParamInt$new("k", default = 7L, lower = 1L, tags = "train"),
@@ -42,12 +38,14 @@ LearnerClassifKKNN = R6Class("LearnerClassifKKNN", inherit = LearnerClassif,
         predict_types = c("response", "prob"),
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
         properties = c("twoclass", "multiclass"),
-        packages = c("withr", "kknn"),
+        packages = "kknn",
         man = "mlr3learners::mlr_learners_classif.kknn"
       )
-    },
+    }
+  ),
 
-    train_internal = function(task) {
+  private = list(
+    .train = function(task) {
       list(
         formula = task$formula(),
         data = task$data(),
@@ -55,11 +53,11 @@ LearnerClassifKKNN = R6Class("LearnerClassifKKNN", inherit = LearnerClassif,
       )
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       model = self$model
       newdata = task$data(cols = task$feature_names)
 
-      withr::with_package("kknn", { # https://github.com/KlausVigo/kknn/issues/16
+      with_package("kknn", { # https://github.com/KlausVigo/kknn/issues/16
         p = invoke(kknn::kknn, formula = model$formula, train = model$data, test = newdata, .args = model$pars)
       })