+++ R/LearnerClassifLogReg.R
@@ -1,26 +1,25 @@
 #' @title Logistic Regression Classification Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_classif.log_reg
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerClassif].
-#'
-#' @section Construction:
-#' ```
-#' LearnerClassifLogReg$new()
-#' mlr3::mlr_learners$get("classif.log_reg")
-#' mlr3::lrn("classif.log_reg")
-#' ```
+#' @name mlr_learners_classif.log_reg
 #'
 #' @description
 #' Classification via logistic regression.
-#' Calls [stats::glm()].
+#' Calls [stats::glm()] with `family` set to `"binomial"`.
+#' Argument `model` is set to `FALSE`.
+#'
+#' @templateVar id classif.log_reg
+#' @template section_dictionary_learner
+#'
+#' @template section_contrasts
 #'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name classif.log_reg
 #' @template example
 LearnerClassifLogReg = R6Class("LearnerClassifLogReg", inherit = LearnerClassif,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       super$initialize(
         id = "classif.log_reg",
@@ -30,18 +29,22 @@ LearnerClassifLogReg = R6Class("LearnerClassifLogReg", inherit = LearnerClassif,
         packages = "stats",
         man = "mlr3learners::mlr_learners_classif.log_reg"
       )
-    },
+    }
+  ),
 
-    train_internal = function(task) {
+  private = list(
+    .train = function(task) {
       pars = self$param_set$get_values(tags = "train")
       if ("weights" %in% task$properties) {
         pars = insert_named(pars, list(weights = task$weights$weight))
       }
 
-      invoke(stats::glm, formula = task$formula(), data = task$data(), family = "binomial", .args = pars)
+      invoke(stats::glm, formula = task$formula(), data = task$data(),
+        family = "binomial", model = FALSE,
+        .args = pars, .opts = opts_default_contrasts)
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       newdata = task$data(cols = task$feature_names)
 
       p = unname(predict(self$model, newdata = newdata, type = "response"))