+++ R/LearnerRegrKKNN.R
@@ -1,19 +1,13 @@
 #' @title k-Nearest-Neighbor Regression Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_regr.kknn
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerRegr].
-#'
-#' @section Construction:
-#' ```
-#' LearnerRegrKKNN$new()
-#' mlr3::mlr_learners$get("regr.kknn")
-#' mlr3::lrn("regr.kknn")
-#' ```
+#' @name mlr_learners_regr.kknn
 #'
 #' @description
 #' k-Nearest-Neighbor regression.
 #' Calls [kknn::kknn()] from package \CRANpkg{kknn}.
+#'
+#' @templateVar id regr.kknn
+#' @template section_dictionary_learner
 #'
 #' @references
 #' \cite{mlr3learners}{hechenbichler_2004}
@@ -24,10 +18,12 @@
 #'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name regr.kknn
 #' @template example
 LearnerRegrKKNN = R6Class("LearnerRegrKKNN", inherit = LearnerRegr,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ParamSet$new(list(
         ParamInt$new("k", default = 7L, lower = 1L, tags = "train"),
@@ -40,12 +36,14 @@ LearnerRegrKKNN = R6Class("LearnerRegrKKNN", inherit = LearnerRegr,
         id = "regr.kknn",
         param_set = ps,
         feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
-        packages = c("withr", "kknn"),
+        packages = "kknn",
         man = "mlr3learners::mlr_learners_regr.kknn"
       )
-    },
+    }
+  ),
 
-    train_internal = function(task) {
+  private = list(
+    .train = function(task) {
       list(
         formula = task$formula(),
         data = task$data(),
@@ -53,11 +51,11 @@ LearnerRegrKKNN = R6Class("LearnerRegrKKNN", inherit = LearnerRegr,
       )
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       model = self$model
       newdata = task$data(cols = task$feature_names)
 
-      withr::with_package("kknn", { # https://github.com/KlausVigo/kknn/issues/16
+      with_package("kknn", { # https://github.com/KlausVigo/kknn/issues/16
         p = invoke(kknn::kknn, formula = model$formula, train = model$data, test = newdata, .args = model$pars)
       })