+++ R/LearnerRegrKM.R
@@ -1,15 +1,6 @@
 #' @title Kriging Regression Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_regr.km
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerRegr].
-#'
-#' @section Construction:
-#' ```
-#' LearnerRegrKM$new()
-#' mlr3::mlr_learners$get("regr.km")
-#' mlr3::lrn("regr.km")
-#' ```
+#' @name mlr_learners_regr.km
 #'
 #' @description
 #' Kriging regression.
@@ -20,15 +11,20 @@
 #'   We recommend a value of `1e-8`.
 #' * The additional hyperparameter `jitter` can be set to add `N(0, [jitter])`-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of `1e-12`.
 #'
+#' @template section_dictionary_learner
+#' @templateVar id regr.km
+#'
 #' @references
 #' \cite{mlr3learners}{roustant_2012}
 #'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name regr.km
 #' @template example
 LearnerRegrKM = R6Class("LearnerRegrKM", inherit = LearnerRegr,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ParamSet$new(list(
         ParamFct$new("covtype", default = "matern5_2", levels = c("gauss", "matern5_2", "matern3_2", "exp", "powexp"), tags = "train"),
@@ -61,14 +57,15 @@ LearnerRegrKM = R6Class("LearnerRegrKM", inherit = LearnerRegr,
         id = "regr.km",
         param_set = ps,
         predict_types = c("response", "se"),
-        feature_types = c("integer", "numeric"),
+        feature_types = c("logical", "integer", "numeric"),
         packages = "DiceKriging",
         man = "mlr3learners::mlr_learners_regr.km"
       )
-    },
+    }
+  ),
 
-    train_internal = function(task) {
-
+  private = list(
+    .train = function(task) {
       pars = self$param_set$get_values(tags = "train")
       data = as.matrix(task$data(cols = task$feature_names))
       truth = task$truth()
@@ -92,7 +89,7 @@ LearnerRegrKM = R6Class("LearnerRegrKM", inherit = LearnerRegr,
       )
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       pars = self$param_set$get_values(tags = "predict")
       newdata = as.matrix(task$data(cols = task$feature_names))