+++ R/LearnerRegrLM.R
@@ -1,47 +1,48 @@
 #' @title Linear Model Regression Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_regr.lm
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerRegr].
-#'
-#' @section Construction:
-#' ```
-#' LearnerRegrLM$new()
-#' mlr3::mlr_learners$get("regr.lm")
-#' mlr3::lrn("regr.lm")
-#' ```
+#' @name mlr_learners_regr.lm
 #'
 #' @description
 #' Ordinary linear regression.
 #' Calls [stats::lm()].
 #'
+#' @templateVar id regr.lm
+#' @template section_dictionary_learner
+#'
+#' @template section_contrasts
+#'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name regr.lm
 #' @template example
 LearnerRegrLM = R6Class("LearnerRegrLM", inherit = LearnerRegr,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       super$initialize(
         id = "regr.lm", ,
         predict_types = c("response", "se"),
-        feature_types = c("integer", "numeric", "factor"),
+        feature_types = c("logical", "integer", "numeric", "factor"),
         properties = "weights",
         packages = "stats",
         man = "mlr3learners::mlr_learners_regr.lm"
       )
-    },
+    }
+  ),
 
-    train_internal = function(task) {
+  private = list(
+    .train = function(task) {
       pars = self$param_set$get_values(tags = "train")
       if ("weights" %in% task$properties) {
         pars = insert_named(pars, list(weights = task$weights$weight))
       }
 
-      invoke(stats::lm, formula = task$formula(), data = task$data(), .args = pars)
+      invoke(stats::lm, formula = task$formula(), data = task$data(),
+        .args = pars, .opts = opts_default_contrasts)
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       newdata = task$data(cols = task$feature_names)
 
       if (self$predict_type == "response") {