+++ R/LearnerRegrXgboost.R
@@ -1,15 +1,6 @@
 #' @title Extreme Gradient Boosting Regression Learner
 #'
-#' @usage NULL
-#' @aliases mlr_learners_regr.xgboost
-#' @format [R6::R6Class()] inheriting from [mlr3::LearnerRegr].
-#'
-#' @section Construction:
-#' ```
-#' LearnerRegrXgboost$new()
-#' mlr3::mlr_learners$get("regr.xgboost")
-#' mlr3::lrn("regr.xgboost")
-#' ```
+#' @name mlr_learners_regr.xgboost
 #'
 #' @description
 #' eXtreme Gradient Boosting regression.
@@ -19,15 +10,20 @@
 #' * Verbosity is reduced by setting `verbose` to `0`.
 #' * Number of boosting iterations `nrounds` is set to `1`.
 #'
+#' @template section_dictionary_learner
+#' @templateVar id regr.xgboost
+#'
 #' @references
 #' \cite{mlr3learners}{chen_2016}
 #'
 #' @export
 #' @template seealso_learner
-#' @templateVar learner_name regr.xgboost
 #' @template example
 LearnerRegrXgboost = R6Class("LearnerRegrXgboost", inherit = LearnerRegr,
   public = list(
+
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ParamSet$new(list(
         ParamFct$new("booster", default = "gbtree", levels = c("gbtree", "gblinear", "dart"), tags = "train"),
@@ -89,9 +85,6 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost", inherit = LearnerRegr,
       ps$add_dep("max_leaves", "grow_policy", CondEqual$new("lossguide"))
       ps$add_dep("max_bin", "tree_method", CondEqual$new("hist"))
       ps$add_dep("sketch_eps", "tree_method", CondEqual$new("approx"))
-      ps$add_dep("lambda", "booster", CondEqual$new("gblinear"))
-      ps$add_dep("lambda_bias", "booster", CondEqual$new("gblinear"))
-      ps$add_dep("alpha", "booster", CondEqual$new("gblinear"))
       ps$add_dep("feature_selector", "booster", CondEqual$new("gblinear"))
       ps$add_dep("top_k", "booster", CondEqual$new("gblinear"))
       ps$add_dep("top_k", "feature_selector", CondAnyOf$new(c("greedy", "thrifty")))
@@ -102,14 +95,32 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost", inherit = LearnerRegr,
       super$initialize(
         id = "regr.xgboost",
         param_set = ps,
-        feature_types = c("integer", "numeric"),
+        feature_types = c("logical", "integer", "numeric"),
         properties = c("weights", "missings", "importance"),
         packages = "xgboost",
         man = "mlr3learners::mlr_learners_regr.xgboost"
       )
     },
 
-    train_internal = function(task) {
+    #' @description
+    #' The importance scores are calculated with [xgboost::xgb.importance()].
+    #'
+    #' @return Named `numeric()`.
+    importance = function() {
+      if (is.null(self$model)) {
+        stopf("No model stored")
+      }
+
+      imp = xgboost::xgb.importance(
+        feature_names = self$model$features,
+        model = self$model
+      )
+      set_names(imp$Gain, imp$Feature)
+    }
+  ),
+
+  private = list(
+    .train = function(task) {
 
       pars = self$param_set$get_values(tags = "train")
 
@@ -132,7 +143,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost", inherit = LearnerRegr,
       invoke(xgboost::xgb.train, data = data, .args = pars)
     },
 
-    predict_internal = function(task) {
+    .predict = function(task) {
       pars = self$param_set$get_values(tags = "predict")
       model = self$model
       newdata = data.matrix(task$data(cols = task$feature_names))
@@ -140,18 +151,6 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost", inherit = LearnerRegr,
       response = invoke(predict, model, newdata = newdata, .args = pars)
 
       PredictionRegr$new(task = task, response = response)
-    },
-
-    importance = function() {
-      if (is.null(self$model)) {
-        stopf("No model stored")
-      }
-
-      imp = xgboost::xgb.importance(
-        feature_names = self$model$features,
-        model = self$model
-      )
-      set_names(imp$Gain, imp$Feature)
     }
   )
 )