+++ NEWS.md
@@ -1,3 +1,13 @@
+# mlr3learners 0.1.7
+
+* Added learner `classif.multinom` from package `nnet`.
+* Learners `regr.lm` and `classif.log_reg` now ignore the global option
+  `"contrasts"`.
+
+# mlr3learners 0.1.6
+
+* Added missing feature type `logical()` to multiple learners.
+
 # mlr3learners 0.1.5
 
 * Added parameter and parameter dependencies to `regr.glmnet`, `regr.km`,